refactor page tests to match changes
